refactor: 💡 line message api
