--- src/apis/README.md
@@ -1,4 +1,0 @@
-# NOTES
-
-- `"use server";`を入れると外部に公開されてしまうため、入力不可。
-- クライアントサイドからのアクセスを禁止したいため、`"import "server-only";`は入力必須。
