+++ src/features/auth/actions/change-role.ts
@@ -1,12 +1,12 @@
 "use server";
 import "server-only";
-import { sendLineNotifyMessage } from "@/apis/line-notify/fetch-message";
 import { SUCCESS_MESSAGES } from "@/constants";
 import { NotAllowedError } from "@/error-classes";
 import { wrapServerSideErrorForClient } from "@/error-wrapper";
 import { getSelfRole } from "@/features/auth/utils/get-session";
 import prisma from "@/prisma";
 import type { ServerAction } from "@/types";
+import { sendLineNotifyMessage } from "@/utils/fetch-message";
 import { formatUpdateRoleMessage } from "@/utils/format-for-line";
 import type { Role } from "@prisma/client";
