+++ src/features/auth/actions/sign-out.tsx
@@ -1,10 +1,10 @@
 "use server";
 import "server-only";
-import { sendLineNotifyMessage } from "@/apis/line-notify/fetch-message";
 import { SUCCESS_MESSAGES } from "@/constants";
 import { wrapServerSideErrorForClient } from "@/error-wrapper";
 import { signOut as NextAuthSignOut } from "@/features/auth/utils/auth";
 import type { ServerAction } from "@/types";
+import { sendLineNotifyMessage } from "@/utils/fetch-message";
 
 type SignOutState = ServerAction<undefined>;
