+++ src/features/profile/actions/change-profile.ts
@@ -1,12 +1,12 @@
 "use server";
 import "server-only";
-import { sendLineNotifyMessage } from "@/apis/line-notify/fetch-message";
 import { SUCCESS_MESSAGES } from "@/constants";
 import { wrapServerSideErrorForClient } from "@/error-wrapper";
 import { getUserId } from "@/features/auth/utils/get-session";
 import type { ProfileSchema } from "@/features/profile/schemas/profile-schema";
 import prisma from "@/prisma";
 import type { ServerAction } from "@/types";
+import { sendLineNotifyMessage } from "@/utils/fetch-message";
 import { formatUpsertProfileMessage } from "@/utils/format-for-line";
 
 export async function changeProfile(
