+++ src/features/profile/actions/change-scope.ts
@@ -1,11 +1,11 @@
 "use server";
 import "server-only";
-import { sendLineNotifyMessage } from "@/apis/line-notify/fetch-message";
 import { SUCCESS_MESSAGES } from "@/constants";
 import { wrapServerSideErrorForClient } from "@/error-wrapper";
 import { getSelfRole, getUserId } from "@/features/auth/utils/get-session";
 import prisma from "@/prisma";
 import type { ServerAction } from "@/types";
+import { sendLineNotifyMessage } from "@/utils/fetch-message";
 import { formatUpdateScopeMessage } from "@/utils/format-for-line";
 
 export async function changeScope(
